+++ gatsby-config.ts
@@ -1,7 +1,8 @@
 import type {GatsbyConfig} from "gatsby";
 import {getImage} from "gatsby-plugin-image";
+import * as dotenv from 'dotenv' // see https://github.com/motdotla/dotenv#how-do-i-use-dotenv-with-import
 
-require('dotenv').config({
+dotenv.config({
     path: `.env.${process.env.NODE_ENV}`
 })